fix: three device crashes — PRAGMA, globalThis, QuickJS stack overflow
Three root causes found by direct crash log analysis on BOOX NoteAir5C:

1. PRAGMA journal_mode crash: BOOX SQLite driver rejects execSQL() for
   PRAGMAs that return a result row. Fixed by switching to rawQuery().close().

2. globalThis unavailable in plugin calls: Plugins registered themselves on
   globalThis[PLUGIN_ID] but callFunction looked up that same global. On
   this QuickJS build, globalThis lookups were unreliable during evaluate()
   calls. Fixed by initializing var __plugins__ = {} in Kotlin and having
   all plugins register into __plugins__[PLUGIN_ID] instead.

3. QuickJS stack overflow: Dispatchers.IO threads have ~512KB native stacks;
   QuickJS requires ≥2MB to evaluate JS function calls. Plugin loading
   (defining functions) succeeded but calling them always overflowed.
   Fixed by replacing limitedParallelism(1) with a dedicated thread created
   via Executors.newSingleThreadExecutor with a 4MB stack.

All three QuickJS entry points (initializeRuntime, loadFromPath, callFunction)
are now pinned to the single 4MB quickjs-thread.

+++ apps/notesprout_android/app/src/main/kotlin/com/notesprout/android/plugins/HostApi.kt
@@ -19,7 +19,7 @@ interface IContextApi {
     fun getPluginId(): String
     fun getObjectId(): String
     fun newId(): String
-    fun now(): Long
+    fun now(): Double
     fun getAppVersion(): String
     // Called from each plugin's IIFE at load time so the registry can discover its pluginId.
     fun registerPlugin(pluginId: String)
@@ -61,7 +61,7 @@ class ContextApi : IContextApi {
     override fun getPluginId(): String = currentPluginId
     override fun getObjectId(): String = currentObjectId
     override fun newId(): String = UUID.randomUUID().toString()
-    override fun now(): Long = System.currentTimeMillis()
+    override fun now(): Double = System.currentTimeMillis().toDouble()
     override fun getAppVersion(): String = BuildConfig.VERSION_NAME
 
     override fun registerPlugin(pluginId: String) {
